user-athentication-edit-delete

+++ src/components/homepage/BlogCard.jsx
@@ -1,13 +1,16 @@
 import React, { useState } from "react";
 import "../homepage/homepage.css";
+import { useAuth } from "../../context/AuthContext";
 
 const BlogCard = ({
   id,
   title: initialTitle,
   content: initialContent,
+  blogAuthorId,
   onDelete,
   onUpdate,
 }) => {
+  const { authorId } = useAuth();
   const [title, setTitle] = useState(initialTitle);
   const [content, setContent] = useState(initialContent);
   const [isEditing, setIsEditing] = useState(false);
@@ -17,14 +20,16 @@ const BlogCard = ({
     const confirmDelete = window.confirm(
       "Are you sure you want to delete this blog?"
     );
-    if (confirmDelete) {
+    if (confirmDelete && blogAuthorId === authorId) {
       onDelete(id);
     }
   };
 
   const handleUpdate = () => {
     setIsEditing(false);
-    onUpdate(id, title, content);
+    if (blogAuthorId === authorId) {
+      onUpdate(id, title, content);
+    }
   };
 
   const toggleFullContent = () => {
@@ -32,71 +37,73 @@ const BlogCard = ({
   };
 
   return (
-    <div className="blogcard-main">
-      <div className="flex mb-4">
-        <div className="w-1/3 mr-4">
-          <img
-            src={`https://picsum.photos/id/${Math.ceil(
-              Math.random() * 200
-            )}/200/250`}
-            alt="Random Image"
-            className="rounded-lg"
-          />
-        </div>
-        <div className="w-2/3">
-          <div className="mb-2 text-center">
-            {isEditing ? (
-              <input
-                type="text"
-                value={title}
-                onChange={(e) => setTitle(e.target.value)}
-                className="blogcard-text-edit2"
-              />
-            ) : (
-              <h2 className="text-2xl font-bold">{title}</h2>
-            )}
+    <div>
+      <div className="blogcard-main">
+        <div className="flex mb-4">
+          <div className="w-1/3 mr-4">
+            <img
+              src={`https://picsum.photos/id/${Math.ceil(
+                Math.random() * 200
+              )}/200/250`}
+              alt="Random Image"
+              className="rounded-lg"
+            />
           </div>
-          <div className="mb-2">
-            {isEditing ? (
-              <textarea
-                value={content}
-                onChange={(e) => setContent(e.target.value)}
-                className="blogcard-text-edit"
-                rows="5"
-              />
-            ) : (
-              <p
-                className={`text-lg font-semibold ${
-                  showFullContent ? "" : "truncate"
-                }`}
-              >
-                {content}
-              </p>
-            )}
+          <div className="w-2/3">
+            <div className="mb-2 text-center">
+              {isEditing ? (
+                <input
+                  type="text"
+                  value={title}
+                  onChange={(e) => setTitle(e.target.value)}
+                  className="blogcard-text-edit2"
+                />
+              ) : (
+                <h2 className="text-2xl font-bold">{title}</h2>
+              )}
+            </div>
+            <div className="mb-2">
+              {isEditing ? (
+                <textarea
+                  value={content}
+                  onChange={(e) => setContent(e.target.value)}
+                  className="blogcard-text-edit"
+                  rows="5"
+                />
+              ) : (
+                <p
+                  className={`text-lg font-semibold ${
+                    showFullContent ? "" : "truncate"
+                  }`}
+                >
+                  {content}
+                </p>
+              )}
+            </div>
           </div>
         </div>
-      </div>
-      <div className="flex justify-between mt-auto">
-        {!isEditing && (
-          <button onClick={toggleFullContent} className="text-blue-500">
-            {showFullContent ? "Read Less" : "Read More"}
-          </button>
-        )}
-        <div>
-          <button
-            onClick={() => setIsEditing(!isEditing)}
-            className="blogcard-edit-btn "
-          >
-            {isEditing ? "Cancel" : "Edit"}
-          </button>
-          {isEditing && (
-            <button onClick={handleUpdate} className="blogcard-update-btn">
-              Update
+        <div className="flex justify-between mt-auto">
+          {!isEditing && (
+            <button onClick={toggleFullContent} className="text-blue-500">
+              {showFullContent ? "Read Less" : "Read More"}
             </button>
           )}
-          <button onClick={handleDelete} className="blogcard-delete-btn">
-            Delete
-          </button>
+          <div>
+            <button
+              onClick={() => setIsEditing(!isEditing)}
+              className="blogcard-edit-btn "
+            >
+              {isEditing ? "Cancel" : "Edit"}
+            </button>
+            {isEditing && (
+              <button onClick={handleUpdate} className="blogcard-update-btn">
+                Update
+              </button>
+            )}
+            <button onClick={handleDelete} className="blogcard-delete-btn">
+              Delete
+            </button>
+          </div>
         </div>
       </div>
     </div>
